Avance desarrollo task #212 - quedan algunos puntos pendientes

+++ src/gc.infraestructura/Dtos/Almacen/Tr/Transferencia/TRNuevaAutSucursalDto.cs
@@ -6,7 +6,7 @@ namespace gc.infraestructura.Dtos.Almacen.Tr.Transferencia
 		public int orden { get; set; } = 0;
 		public string adm_id { get; set; } = string.Empty;
 		public string adm_nombre { get; set; } = string.Empty;
-		public int pallet_aprox { get; set; } = 0;
+		public decimal pallet_aprox { get; set; } = 0;
 		public string nota { get; set; } = string.Empty;
 		public int aut_a_generar { get; set; } = 0;
     }
